+++ src/extension.ts
@@ -1,5 +1,6 @@
 import * as vscode from 'vscode'
 
+import { makeActivationCommand } from './activationcommand'
 import { makeAnnotations } from './annotations'
 import { makeApiSearch } from './apisearch'
 import { makeClient } from './client'
@@ -90,17 +91,4 @@ function restart(output: vscode.OutputChannel, context: vscode.ExtensionContext)
     output.appendLine(alloglot.ui.readyToRestart)
     activate(context)
   })
-}
-
-function makeActivationCommand(parentOutput: IHierarchicalOutputChannel, command: string | undefined, reveal: boolean | undefined): vscode.Disposable {
-  if (!command) return vscode.Disposable.from()
-  const basedir = vscode.workspace.workspaceFolders?.[0].uri
-  const output = parentOutput.split()
-  reveal && output.show(true)
-
-  const proc = AsyncProcess.make({ output, command, basedir }, () => {
-    parentOutput.appendLine(alloglot.ui.activateCommandDone(command))
-  })
-
-  return vscode.Disposable.from(proc, output)
 }
